# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meal Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a new account with email, display name, and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created user session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Email already taken
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password reset email (always returns success to avoid enumeration)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: Reset email sent (or silently skipped if email not found)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password using a reset token from email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid, expired, or already-used token
  /auth/password:
    patch:
      operationId: changePassword
      tags: [auth]
      summary: Change password for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Current password incorrect or not authenticated

  # ── MEALS ─────────────────────────────────────────────────────────────────
...
  /billing/checkout:
    post:
      operationId: startCheckout
      tags: [billing]
      summary: Create a Stripe Checkout session for a subscription tier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartCheckoutBody"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
        "503":
          description: Stripe not configured
  /billing/portal:
    post:
      operationId: openBillingPortal
      tags: [billing]
      summary: Create a Stripe Billing Portal session for the current subscriber
      responses:
        "200":
          description: Portal session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
        "400":
          description: No active subscription
...
components:
  schemas:
...
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        displayName:
          type: string
        tier:
          type: string
          description: "Subscription tier: free | pro | pro_ai"
      required: [id, email, displayName, tier]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string
        password:
          type: string
          minLength: 6
      required: [email, displayName, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    ResetPasswordBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 6
      required: [token, password]
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
      required: [currentPassword, newPassword]
...
    StartCheckoutBody:
      type: object
      properties:
        tier:
          type: string
          description: "pro | pro_ai"
      required: [tier]
    CheckoutResult:
      type: object
      properties:
        url:
          type: string
          description: Stripe-hosted redirect URL
      required: [url]
...